+++ src/lib/string/yapl_string.h
@@ -12,9 +12,9 @@
 #include "../../vm/yapl_vm.h"
 
 /* String operations */
-uint32_t hashString(const char *key, uint64_t length);
+uint32_t hashString(const char *key, int length);
 ObjString *makeString(VM *vm, int length);
-ObjString *copyString(VM *vm, const char *chars, uint64_t length);
+ObjString *copyString(VM *vm, const char *chars, int length);
 ObjString *concatStrings(VM *vm, ObjString *str1, ObjString *str2);
 
 #endif // YAPL_STRING_H
